Src/adf_monitor.py
from azure.identity import DefaultAzureCredential
from azure.mgmt.datafactory import DataFactoryManagementClient
import logging

class ADFMonitor:
    def __init__(self, subscription_id: str, resource_group: str, factory_name: str):
        """Initialize Azure Data Factory monitor"""
        self.subscription_id = subscription_id
        self.resource_group = resource_group
        self.factory_name = factory_name
        self.credential = DefaultAzureCredential()
        self.logger = logging.getLogger(__name__)
        
        self.client = DataFactoryManagementClient(
            credential=self.credential,
            subscription_id=subscription_id
        )
        self.logger.info("ADFMonitor initialized")

    def check_pipeline_status(self):
        """Example monitoring method"""
        pipelines = self.client.pipelines.list_by_factory(
            self.resource_group,
            self.factory_name
        )
        # ... monitoring logic