requirements.txt
gradio==3.50.2
pyyaml==6.0.1
ffmpeg-python==0.2.0
azure-identity==1.14.0
azure-mgmt-datafactory==2.0.0
smtplib==0.0.1
PyGithub==1.59.0
azure-identity>=1.12.0
azure-mgmt-datafactory>=2.0.0
python-dotenv>=0.19.0
playwright>=1.28.0
gradio>=3.0